+++ src/iotvm-eventengine/src/main/java/org/softwareforce/iotvm/eventengine/persistence/IBOPersistenceServiceImpl.java
@@ -15,7 +15,6 @@ import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.Sorts;
 import com.mongodb.client.result.InsertOneResult;
-
 import java.time.Instant;
 import java.util.Optional;
 import javax.annotation.Nullable;